Indicate whether the current page has been bookmarked
I'm using an Event Page (background.js) to listen for changes to tabs: I'm
listening for when tabs become active and when their URL changes. These two
events appear to cover all the situations where I want to send a request to
roosmarks to determine whether I've bookmarked the current URL. *NOTE*
Although I'm making a request to roosmarks for each tab change/update, I still
make a second request when you actually click on the roosmarks extension
button. It's possible I'll be able to store the results from the initial
lookup so that I can avoid the second request, but I'll look at that
separately.

I've had to copy the `showBookmark` function to background.js as I don't yet
know how to share code between javascript files in the extension.

I'm specifying the tabId when using `setBadgeText` so that the text resets
when changing tabs.

I've updated both `showBookmark` functions such that they check for a 200
status code. I still get an annoying 404 warning in the developer console when
a bookmark isn't found but I'll investigate that separately.